Big Count: first pass for coll framework bc
This commit adds only those functions which make use of C integer promotion.
So none of the 'v,w' and reduce_scatter related methods are added in this PR.

Related to #12336
Pieces of #12478 were taken out to make this PR.

+++ ompi/mca/coll/accelerator/coll_accelerator_reduce_scatter_block.c
@@ -32,7 +32,7 @@
  *     up at some point)
  */
 int
-mca_coll_accelerator_reduce_scatter_block(const void *sbuf, void *rbuf, int rcount,
+mca_coll_accelerator_reduce_scatter_block(const void *sbuf, void *rbuf, size_t rcount,
                                    struct ompi_datatype_t *dtype,
                                    struct ompi_op_t *op,
                                    struct ompi_communicator_t *comm,
