+++ ompi/mca/coll/demo/coll_demo_allreduce.c
@@ -34,7 +34,7 @@
  *	Accepts:	- same as MPI_Allreduce()
  *	Returns:	- MPI_SUCCESS or error code
  */
-int mca_coll_demo_allreduce_intra(const void *sbuf, void *rbuf, int count,
+int mca_coll_demo_allreduce_intra(const void *sbuf, void *rbuf, size_t count,
                                   struct ompi_datatype_t *dtype,
                                   struct ompi_op_t *op,
                                   struct ompi_communicator_t *comm,
@@ -55,7 +55,7 @@ int mca_coll_demo_allreduce_intra(const void *sbuf, void *rbuf, int count,
  *	Accepts:	- same as MPI_Allreduce()
  *	Returns:	- MPI_SUCCESS or error code
  */
-int mca_coll_demo_allreduce_inter(const void *sbuf, void *rbuf, int count,
+int mca_coll_demo_allreduce_inter(const void *sbuf, void *rbuf, size_t count,
                                   struct ompi_datatype_t *dtype,
                                   struct ompi_op_t *op,
                                   struct ompi_communicator_t *comm,
