+++ ompi/mca/coll/demo/coll_demo_gather.c
@@ -33,9 +33,9 @@
  *	Accepts:	- same arguments as MPI_Gather()
  *	Returns:	- MPI_SUCCESS or error code
  */
-int mca_coll_demo_gather_intra(const void *sbuf, int scount,
+int mca_coll_demo_gather_intra(const void *sbuf, size_t scount,
                                struct ompi_datatype_t *sdtype,
-                               void *rbuf, int rcount,
+                               void *rbuf, size_t rcount,
                                struct ompi_datatype_t *rdtype,
                                int root, struct ompi_communicator_t *comm,
                                mca_coll_base_module_t *module)
@@ -56,9 +56,9 @@ int mca_coll_demo_gather_intra(const void *sbuf, int scount,
  *	Accepts:	- same arguments as MPI_Gather()
  *	Returns:	- MPI_SUCCESS or error code
  */
-int mca_coll_demo_gather_inter(const void *sbuf, int scount,
+int mca_coll_demo_gather_inter(const void *sbuf, size_t scount,
                                struct ompi_datatype_t *sdtype,
-                               void *rbuf, int rcount,
+                               void *rbuf, size_t rcount,
                                struct ompi_datatype_t *rdtype,
                                int root, struct ompi_communicator_t *comm,
                                mca_coll_base_module_t *module)
