+++ ompi/mca/coll/demo/coll_demo_reduce.c
@@ -34,7 +34,7 @@
  *	Accepts:	- same as MPI_Reduce()
  *	Returns:	- MPI_SUCCESS or error code
  */
-int mca_coll_demo_reduce_intra(const void *sbuf, void *rbuf, int count,
+int mca_coll_demo_reduce_intra(const void *sbuf, void *rbuf, size_t count,
                                struct ompi_datatype_t *dtype,
                                struct ompi_op_t *op,
                                int root, struct ompi_communicator_t *comm,
@@ -55,7 +55,7 @@ int mca_coll_demo_reduce_intra(const void *sbuf, void *rbuf, int count,
  *	Accepts:	- same as MPI_Reduce()
  *	Returns:	- MPI_SUCCESS or error code
  */
-int mca_coll_demo_reduce_inter(const void *sbuf, void *rbuf, int count,
+int mca_coll_demo_reduce_inter(const void *sbuf, void *rbuf, size_t count,
                                struct ompi_datatype_t *dtype,
                                struct ompi_op_t *op,
                                int root, struct ompi_communicator_t *comm,
