+++ ompi/mca/coll/han/coll_han_bcast.c
@@ -65,7 +65,7 @@ mca_coll_han_set_bcast_args(mca_coll_han_bcast_args_t * args, mca_coll_task_t * 
  */
 int
 mca_coll_han_bcast_intra(void *buf,
-                         int count,
+                         size_t count,
                          struct ompi_datatype_t *dtype,
                          int root,
                          struct ompi_communicator_t *comm, mca_coll_base_module_t * module)
@@ -113,7 +113,7 @@ mca_coll_han_bcast_intra(void *buf,
 
     int num_segments = (count + seg_count - 1) / seg_count;
     OPAL_OUTPUT_VERBOSE((20, mca_coll_han_component.han_output,
-                         "In HAN seg_count %d count %d num_seg %d\n",
+                         "In HAN seg_count %d count %zu num_seg %d\n",
                          seg_count, count, num_segments));
 
     int *vranks = han_module->cached_vranks;
@@ -224,7 +224,7 @@ int mca_coll_han_bcast_t1_task(void *task_args)
  */
 int
 mca_coll_han_bcast_intra_simple(void *buf,
-                                int count,
+                                size_t count,
                                 struct ompi_datatype_t *dtype,
                                 int root,
                                 struct ompi_communicator_t *comm,
