+++ ompi/mca/coll/self/coll_self_allgather.c
@@ -32,9 +32,9 @@
  *	Accepts:	- same as MPI_Allgather()
  *	Returns:	- MPI_SUCCESS, or error code
  */
-int mca_coll_self_allgather_intra(const void *sbuf, int scount,
+int mca_coll_self_allgather_intra(const void *sbuf, size_t scount,
                                   struct ompi_datatype_t *sdtype, void *rbuf,
-                                  int rcount, struct ompi_datatype_t *rdtype,
+                                  size_t rcount, struct ompi_datatype_t *rdtype,
                                   struct ompi_communicator_t *comm,
                                   mca_coll_base_module_t *module)
 {
