+++ ompi/mca/coll/self/coll_self_reduce.c
@@ -32,7 +32,7 @@
  *	Accepts:	- same as MPI_Reduce()
  *	Returns:	- MPI_SUCCESS or error code
  */
-int mca_coll_self_reduce_intra(const void *sbuf, void *rbuf, int count,
+int mca_coll_self_reduce_intra(const void *sbuf, void *rbuf, size_t count,
                                struct ompi_datatype_t *dtype,
                                struct ompi_op_t *op,
                                int root, struct ompi_communicator_t *comm,
