+++ ompi/mca/coll/tuned/coll_tuned_reduce_scatter_block_decision.c
@@ -115,7 +115,7 @@ int ompi_coll_tuned_reduce_scatter_block_intra_check_forced_init (coll_tuned_for
 }
 
 int ompi_coll_tuned_reduce_scatter_block_intra_do_this(const void *sbuf, void *rbuf,
-                                                       int rcount,
+                                                       size_t rcount,
                                                        struct ompi_datatype_t *dtype,
                                                        struct ompi_op_t *op,
                                                        struct ompi_communicator_t *comm,
